Character Sprites Fixed and New scene Position fixed
Fixed Sprites so the gun looks the same for all character sprites.

When you enter a new area you usually are in a fixed position. Now that is different you are placed in the proper position from where your coming from in the pervious room.
//If Entering a Left scene then (X = -X And (Y  -Y) = itself), (Right -X = X and (Y  -Y) = itself), (Up (X  -X) = itself and Y = -Y), (Down (X  -X) = itself and -Y = Y)

also added first particle effect for fun. It's a healing effect. It's in the character selection screen and the first main room.